+++ typescript/pragma-scripts/package.json
@@ -18,7 +18,7 @@
     "pragma-utils": "file:../pragma-utils"
   },
   "scripts": {
-    "pragma:update_feed": "bun run src/pragmaSol/update_feed.ts",
+    "pragma:update_feed": "bun run src/pragma/update_feed.ts",
     "dispatcher:add_feed": "bun run src/dispatcher/add_feed.ts",
     "dispatcher:get_all_feeds": "bun run src/dispatcher/get_all_feeds.ts",
     "dispatcher:dispatch": "bun run src/dispatcher/dispatch.ts"
