+++ typescript/pragma-scripts/src/dispatcher/add_feeds.ts
@@ -1,6 +1,5 @@
-import fs from "fs";
 import { Command, type OptionValues } from "commander";
-import type { Account, Contract } from "starknet";
+import type { Contract } from "starknet";
 import {
   buildAccount,
   Deployer,
@@ -12,11 +11,11 @@ import {
 function parseCommandLineArguments(): OptionValues {
   const program = new Command();
   program
-    .name("remove-feeds")
-    .description("CLI to remove multiple Pragma data feeds")
+    .name("add-feeds")
+    .description("CLI to update multiple Pragma data feeds")
     .requiredOption(
       "--feed-ids <feed_ids...>",
-      "IDs of the data feeds to remove",
+      "IDs of the data feeds to update",
     )
     .requiredOption(
       "--chain <chain_name>",
@@ -34,24 +33,25 @@ function parseCommandLineArguments(): OptionValues {
   return options;
 }
 
-async function removeFeed(
+async function addFeed(
   pragmaDispatcher: Contract,
   account: Deployer,
   feedId: string,
 ) {
   try {
-    console.log(`Removing feed ${feedId}`);
-    const invoke = await pragmaDispatcher.invoke("remove_feed", [feedId]);
+    console.log(`\t ⏳ Adding feed ${feedId}...`);
+    const invoke = await pragmaDispatcher.invoke("add_feed", [feedId]);
     await account.waitForTransaction(invoke.transaction_hash);
 
     const receipt = await account.getTransactionReceipt(
       invoke.transaction_hash,
     );
     await ensureSuccess(receipt, account.provider);
-    console.log(`Successfully removed feed ${feedId}`);
+    console.log(`\t 🧩 Successfully added feed ${feedId}`);
   } catch (error) {
-    console.error(`Error removing feed ${feedId}:`, error);
+    console.error(`\t⛔ Error adding feed ${feedId}:`, error);
   }
+  console.log("✅ All feeds added!");
 }
 
 async function main() {
@@ -65,12 +65,12 @@ async function main() {
   const feedIds = options.feedIds;
   const account = await buildAccount(options.chain);
   console.log(
-    `👉 Removing feeds for contract ${publisherRegistryAddress} on chain ${options.chain}`,
+    `👉 Adding feeds for contract ${publisherRegistryAddress} on chain ${options.chain}`,
   );
 
   const pragmaDispatcher = await account.loadContract(publisherRegistryAddress);
   for (const feedId of feedIds) {
-    await removeFeed(pragmaDispatcher, account, feedId);
+    await addFeed(pragmaDispatcher, account, feedId);
   }
 }
