+++ typescript/pragma-scripts/src/dispatcher/remove_feeds.ts
@@ -1,5 +1,6 @@
+import fs from "fs";
 import { Command, type OptionValues } from "commander";
-import type { Contract } from "starknet";
+import type { Account, Contract } from "starknet";
 import {
   buildAccount,
   Deployer,
@@ -11,11 +12,11 @@ import {
 function parseCommandLineArguments(): OptionValues {
   const program = new Command();
   program
-    .name("add-feeds")
-    .description("CLI to update multiple Pragma data feeds")
+    .name("remove-feeds")
+    .description("CLI to remove multiple Pragma data feeds")
     .requiredOption(
       "--feed-ids <feed_ids...>",
-      "IDs of the data feeds to update",
+      "IDs of the data feeds to remove",
     )
     .requiredOption(
       "--chain <chain_name>",
@@ -33,25 +34,25 @@ function parseCommandLineArguments(): OptionValues {
   return options;
 }
 
-async function addFeed(
+async function removeFeed(
   pragmaDispatcher: Contract,
   account: Deployer,
   feedId: string,
 ) {
   try {
-    console.log(`\t ⏳ Adding feed ${feedId}...`);
-    const invoke = await pragmaDispatcher.invoke("add_feed", [feedId]);
+    console.log(`\t ⏳ Removing feed ${feedId}...`);
+    const invoke = await pragmaDispatcher.invoke("remove_feed", [feedId]);
     await account.waitForTransaction(invoke.transaction_hash);
 
     const receipt = await account.getTransactionReceipt(
       invoke.transaction_hash,
     );
     await ensureSuccess(receipt, account.provider);
-    console.log(`\t 🧩 Successfully added feed ${feedId}`);
+    console.log(`\t 🧩 Successfully removed feed ${feedId}`);
   } catch (error) {
-    console.error(`Error adding feed ${feedId}:`, error);
+    console.error(`Error removing feed ${feedId}:`, error);
   }
-  console.log("✅ All feeds added!");
+  console.log("✅ All feeds removed!");
 }
 
 async function main() {
@@ -65,12 +66,12 @@ async function main() {
   const feedIds = options.feedIds;
   const account = await buildAccount(options.chain);
   console.log(
-    `👉 Adding feeds for contract ${publisherRegistryAddress} on chain ${options.chain}`,
+    `👉 Removing feeds for contract ${publisherRegistryAddress} on chain ${options.chain}`,
   );
 
   const pragmaDispatcher = await account.loadContract(publisherRegistryAddress);
   for (const feedId of feedIds) {
-    await addFeed(pragmaDispatcher, account, feedId);
+    await removeFeed(pragmaDispatcher, account, feedId);
   }
 }
